machine: writerKernelModules: Add support for modules.builtin
In recent kernels, ArchLinux no longer compiles some of the virtio
drivers as modules; they are compiled built-in in the kernel image
instead. This seems to be also true for Ubuntu.

This commit adds support for the `modules.builtin` file. We now read
this file, and skip the builtin modules.

Note that we can't assume that the file `modules.builtin` exist. On
Debian, modules.builtin doesn't exist at least on the 4.19 kernels it
does however exist on the 5.3 packaging

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>